Backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
email-validator==2.1.0.post1
pdfplumber

